Serralongue now runs on x64 again (patched DockPanelSuite problem: http://sourceforge.net/tracker/index.php?func=detail&aid=1438642&group_id=110642&atid=659401)
git-svn-id: svn://svn.sharpdevelop.net/sharpdevelop/trunk@1335 1ccf3a8d-04fe-1044-b7c0-cef0b8235c61

+++ src/Libraries/DockPanel_Src/WinFormsUI/Win32/Structs.cs
@@ -136,10 +136,10 @@ namespace WeifenLuo.WinFormsUI.Win32
 	[StructLayout(LayoutKind.Sequential)]
 	internal struct CREATESTRUCT 
 	{
-		public int lpCreateParams;
-		public int hInstance;
-		public int hMenu;
-		public int hWndParent;
+        public IntPtr lpCreateParams;
+        public IntPtr hInstance;
+        public IntPtr hMenu;
+        public IntPtr hWndParent;
 		public int cy;
 		public int cx;
 		public int y;
@@ -155,20 +155,20 @@ namespace WeifenLuo.WinFormsUI.Win32
 	{
 		public string szClass;
 		public string szTitle;
-		public int hOwner;
+        public IntPtr hOwner;
 		public int x;
 		public int y;
 		public int cx;
 		public int cy;
 		public int style;
-		public int lParam;
+		public IntPtr lParam;
 	}
 
 	[StructLayout(LayoutKind.Sequential)]
 	internal struct WINDOWPOS 
 	{
-		public int hwnd;
-		public int hWndInsertAfter;
+		public IntPtr hwnd;
+        public IntPtr hWndInsertAfter;
 		public int x;
 		public int y;
 		public int cx;
